removed old tests and updated to get username from jwt
