+++ analytics/utils/__init__.py
@@ -2,3 +2,4 @@
 
 from .db import *
 from .calculations import *
+from .get_user import *
